du_cmm: Fix release for Yocto v3.21.0
When the HDMI cable is pulled out during suspend/resume,
the DU clock may stop before suspending the CMM.
If the CMM register is accessed while DU stopped, it hangs
due to the H/W specifications.
To avoid this, move cmm register backup before stopping DU.

And the value of the CMM register is saved just
before the DU stops, so that it can be set during resuming.

Signed-off-by: Koji Matsuoka <koji.matsuoka.xm@renesas.com>